[RISC-V][PR target/121531] Cover missing insn types in p400 and p600 scheduler models
So the usual problems, DFAs without full coverage.  I took the output of Kito's
checker and use that to construct a dummy reservation for the p400 and p600
sifive models.

Tested on riscv32-elf and riscv64-elf with no regressions.

Pushing to the trunk once pre-commit CI gives the green light.

	PR target/121531
gcc/
	* config/riscv/sifive-p400.md (sifive_p400_unknown): New reservation.
	* config/riscv/sifive-p600.md (sifive_p600_unkonwn): Likewise.

gcc/testsuite/
	* gcc.target/riscv/pr121531.c: New test.