testsuite: Fix asm-hard-reg-error-3.c for arm [PR121511]
This test is about register pairs.  On arm a long long is accepted in
thumb mode in any register 0-6 whereas in arm mode this is restricted to
even register pairs.  Thus, in order to trigger the error even if gcc is
configured with --with-mode=thumb, add option -marm.

gcc/testsuite/ChangeLog:

	PR target/121511
	* gcc.dg/asm-hard-reg-error-3.c: Add additional option -marm for
	target arm.