[PR target/119275][RISC-V] Avoid calling gen_lowpart in cases where it would ICE
So this is a minor bug in the riscv move expanders.  It has a special cases for
extraction from vector objects which makes assumptions that it can use
gen_lowpart unconditionally.  That's not always the case.

We can just bypass that special code for cases where we can't use gen_lowpart
and let the more generic code run. If gen_lowpart_common indicates we've got a
case that can't be handled we just bypass the special extraction code.

Tested on riscv64-elf and riscv32-elf.  Waiting for pre-commit CI to do its
thing.

	PR target/119275
gcc/
	* config/riscv/riscv.cc (riscv_legitimize_move): Avoid calling
	gen_lowpart for cases where it'll fail.  Just use standard expander
	paths for those cases.

gcc/testsuite/
	* gcc.target/riscv/pr119275.c: New test.