Generate a call to a .ACCESS_WITH_SIZE for a FAM with counted_by attribute only when it's read from.
Currently, we generate a call to a .ACCESS_WITH_SIZE for a FAM with counted_by
attribute for every component_ref that corresponds to such an object.
Actually, such .ACCESS_WITH_SIZE calls are useless when they are generated
for a written site or an address taken site.

In this patch, we only generate a call to .ACCESS_WITH_SIZE for a FAM with
counted_by attribute when it's a read.

gcc/c/ChangeLog:

	* c-tree.h (handle_counted_by_for_component_ref): New prototype of
	build_component_ref and handle_counted_by_for_component_ref.
	* c-parser.cc (c_parser_postfix_expression): Call the new prototypes
	of build_component_ref and handle_counted_by_for_component_ref,
	update comments.
	* c-typeck.cc (default_function_array_read_conversion): Likewise.
	(convert_lvalue_to_rvalue): Likewise.
	(default_conversion): Likewise.
	(handle_counted_by_p): Update comments.
	(handle_counted_by_for_component_ref): Delete one argument.
	(build_component_ref): Delete one argument. Delete the call to
	handle_counted_by_for_component_ref completely.
	(build_array_ref): Generate call to .ACCESS_WITH_SIZE for array.

gcc/testsuite/ChangeLog:

	* gcc.dg/flex-array-counted-by-2.c: Adjust testing case.