AArch64: Fix scalar costing after removal of vectype from mid-end [PR121536]
commit g:fb59c5719c17a04ecfd58b5e566eccd6d2ac583a stops passing the scalar type
(confusingly named vectype) to the costing hook when doing scalar costing.

As a result, we could no longer distinguish between FPR and GPR scalar stmts.
A later commit also removed STMT_VINFO_VECTYPE from stmt_info.

This leaves the only remaining option to get the type of the original stmt in
the stmt_info.  This patch does this when we're performing scalar costing.

Ideally I'd refactor this a bit because a lot of the hooks just need to know if
it's FP or not, but this seems pointless with the ongoing costing churn.  So for
now this restores our costing.

gcc/ChangeLog:

	PR target/121536
	* config/aarch64/aarch64.cc (aarch64_vector_costs::add_stmt_cost): Set
	vectype from type of lhs of gimple stmt.