rtl-ssa: Fix thinko when adding live-out uses
While testing a later patch, I found that create_degenerate_phi
had an inverted test for bitmap_set_bit.  It was assuming that
the return value was the previous bit value, rather than a
"something changed" value. :(

Also, the call to add_live_out_use shouldn't be conditional
on the DF_LR_OUT operation, since the register could be live-out
because of uses later in the same EBB (which do not require a
live-out use to be added to the rtl-ssa instruction).  Instead,
add_live_out should itself check whether a live-out use already exists.

gcc/
	* rtl-ssa/blocks.cc (function_info::create_degenerate_phi): Fix
	inverted test of bitmap_set_bit.  Call add_live_out_use even
	if the register was previously live-out from the predecessor block.
	Instead...
	(function_info::add_live_out_use): ...check here whether a live-out
	use already exists.